+++ charts/manticoresearch/CHANGELOG.md
@@ -3,3 +3,4 @@
 * Start using domains in replication instead IP 
 * Changed binlog dir (remove data folder from default path)
 * Readiness probe has been removed
+* Added check for max allowed length of worker hostname